* Replace column index/count type (i32) with usize.
* Replace parameter index/count type (i32) with usize.
* Replace row changes/count type (i32) with usize.

+++ rusqlite/src/functions.rs
@@ -216,7 +216,7 @@ impl<'a> Context<'a> {
                 Error::InvalidFunctionParameterType(idx, value.data_type())
             }
                                                   FromSqlError::OutOfRange(i) => {
-                                                      Error::IntegralValueOutOfRange(idx as c_int,
+                                                      Error::IntegralValueOutOfRange(idx,
                                                                                      i)
                                                   }
                                                   FromSqlError::Other(err) => {
